requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==4.2.13
django-allauth==0.63.3
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
MarkupSafe==2.1.5
msgpack==1.0.8
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
redis==5.0.7
requests==2.32.3
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
zope.interface==6.4.post2